+++ src/main/java/it/polimi/db2/db2project/webmodule/controllers/Confirmation.java
@@ -92,7 +92,7 @@ public class Confirmation extends HttpServlet{
                     List<OptionalProduct> optionals = new ArrayList<>();
                     if (request.getParameter("optionals") != null) {
                         for (String optProd : request.getParameterValues("optionals")) {
-                            OptionalProduct optionalProduct = optionalService.findOptionalByID(Long.parseLong(optProd));
+                            OptionalProduct optionalProduct = optionalService.findOptionalByID(Integer.parseInt(optProd));
                             optionals.add(optionalProduct);
                         }
                     }
